now detached node relations before delete

+++ src/main/java/com/adscoop/userservice/handlers/credit/DeleteCreditCardHandler.java
@@ -23,13 +23,18 @@ public class DeleteCreditCardHandler implements Handler {
     @Override
     public void handle(Context ctx) throws Exception {
         String token = ctx.getRequest().getHeaders().get("token");
-        CreditInfo creditInfo = creditInfoService.findByUserToken(token);
-        if (creditInfo != null) {
-            try {
-                creditInfoService.delete(creditInfo);
-            }catch (Exception e){
+        if (ctx.getRequest().getMethod().isDelete()) {
+            CreditInfo creditInfo = creditInfoService.findByUserToken(token);
+            if (creditInfo != null) {
+                try {
+                    creditInfoService.delete(creditInfo);
+                } catch (Exception e) {
 
+                }
             }
+        } else {
+            ctx.next();
         }
+
     }
 }
